refactored the auth and login endpoints and persistance logics from user resources to auth resources
