+++ src/app.module.ts
@@ -2,7 +2,7 @@
  * @Author: Ardrit Krasniqi 
  * @Date: 2021-10-16 23:44:21 
  * @Last Modified by: Ardrit Krasniqi ©
- * @Last Modified time: 2021-12-30 16:19:15
+ * @Last Modified time: 2022-01-04 15:01:32
  */
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
@@ -24,6 +24,7 @@ import { EmailModule } from './email/email.module';
     UsersModule,
     DatabaseModule,
     EmailModule
-  ]
+  ],
+  providers: []
 })
 export class AppModule {}
