+++ src/users/user.entity.ts
@@ -1,9 +1,10 @@
 /*
  * @Author: Ardrit Krasniqi © 
  * @Date: 2022-01-03 15:39:33 
- * @Last Modified by:   Ardrit Krasniqi © 
- * @Last Modified time: 2022-01-03 15:39:33 
+ * @Last Modified by: Ardrit Krasniqi ©
+ * @Last Modified time: 2022-01-04 16:12:17
  */
+import { Exclude } from "class-transformer";
 import { BaseEntity, Column, CreateDateColumn, Entity, PrimaryGeneratedColumn, Unique, UpdateDateColumn } from "typeorm";
 
 @Entity('users')
@@ -25,6 +26,7 @@ export class User extends BaseEntity{
         length: 255,
         nullable: false
     })
+    @Exclude()
     password: string;
 
     @Column({
