+++ src/users/users.module.ts
@@ -1,8 +1,8 @@
 /*
  * @Author: Ardrit Krasniqi © 
  * @Date: 2022-01-03 15:38:31 
- * @Last Modified by:   Ardrit Krasniqi © 
- * @Last Modified time: 2022-01-03 15:38:31 
+ * @Last Modified by: Ardrit Krasniqi ©
+ * @Last Modified time: 2022-01-04 15:04:55
  */
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
